+++ src/components/Driver/DriverSignup.js
@@ -7,6 +7,7 @@ import RegistrationForm from '../RegistrationForm';
 import { mapStateToProps } from '../../redux/actions';
 import SignUpHelper from '../../helpers/SignUpHelper';
 import ErrorMessages from '../layouts/ErrorMessages';
+import imgUploadHelper from '../../helpers/imgUploadHelper';
 
 const DriverSignup = ({ appState, dispatch }) => {
   const history = useHistory();
@@ -21,22 +22,7 @@ const DriverSignup = ({ appState, dispatch }) => {
   };
 
   const onChange = (event) => {
-    const parameter = {
-      ...fields,
-      alert: '',
-    };
-    if (event.target.files && event.target.files[0]) {
-      const img = event.target.files[0];
-      const reader = new FileReader();
-      reader.onloadend = () => {
-        parameter[event.target.id] = reader.result;
-        setFields(parameter);
-      };
-      reader.readAsDataURL(img);
-    } else {
-      parameter[event.target.id] = event.target.value;
-      setFields(parameter);
-    }
+    imgUploadHelper({ event, fields, setFields });
   };
 
   const onClick = (event, Fields, SetFields, State, Dispatch, History, Params) => {
