+++ src/utils/ProfileInfoTable.js
@@ -1,14 +1,14 @@
 import React from 'react';
 
 const ProfileInfoTable = ({ fieldsData }) => {
-  const fields = Object.entries(fieldsData).map((key) => (
-    key[0] !== 'driver_license_image_id'
+  const fields = Object.entries(fieldsData).map(([key, value]) => (
+    key !== 'driver_license_image_id'
     && (
-    <table key={key[0]} className="table table-borderless">
+    <table key={key} className="table table-borderless">
       <thead>
         <tr>
-          <th>{ key[0] }</th>
-          <th>{typeof key[1] === 'boolean' ? key[1].toString() : key[1]}</th>
+          <th>{ key }</th>
+          <th>{typeof value === 'boolean' ? value.toString() : value}</th>
         </tr>
       </thead>
     </table>
